requirements.txt
Flask==0.12
Jinja2==2.9.5
MarkupSafe==0.23
MySQL-python==1.2.5
PyYAML==3.12
REfO==0.13
SPARQLWrapper==1.8.0
Werkzeug==0.11.15
apiai==1.2.3
click==6.7
docopt==0.6.2
isodate==0.5.4
itsdangerous==0.24
nltk==3.2.2
numpy==1.12.1
pymongo==3.4.0
pyparsing==2.2.0
quepy==0.2
rdflib==4.2.2
requests==2.13.0
simplejson==3.10.0
six==1.10.0
wsgiref==0.1.2
